Renamed FLATFILE databasetype to YAML
